API Key Creation & Management
Added functionality for superusers and users to create and manage API keys, with Knox integration for secure key hashing.
